Move preprocessing to proof_system module
